// src/app/views/pages/estadisticas/models/resultadoBase.ts
export interface ResultadoBase {
    cantidad: number;
    contactadas: number;
    id: number;
    invendibles: number;
    llamadas: number;
    nombre: string;
    porcentajeContactadas: number;
    porcentajeInvendibles: number;
    porcentajeVendibles: number;
    porcentajeVendidas: number;
    vendibles: number;
    vendidas: number;
}